extract BlockManagerDecommissionManager into a toplevel class
